requirements.txt
mcp[cli]>=0.1.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
flask>=2.0.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
werkzeug>=2.0.0
langchain>=0.1.0
transformers>=4.30.0
sentence-transformers>=2.2.0
fastapi>=0.100.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
pydantic>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
spacy>=3.7.0
medspacy>=0.2.0
pydicom>=2.4.0
nibabel>=5.1.0
scipy>=1.11.0
biosppy>=1.0.0